APIEngine/requirements.txt
astroid
autopep8
backports.functools-lru-cache
certifi
chardet
configparser
Django
django-cors-headers
django-markdown-deux
djangorestframework
djangorestframework-jwt
enum34
isort
lazy-object-proxy
markdown2
mccabe
Pillow
psycopg2
psycopg2-binary
pycodestyle
PyJWT
pylint
pytz
requests
singledispatch
six
urllib3
wrapt
